server/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
tqdm
pinecone
langchain
langchain-community
langchain-huggingface
langchain-google-genai
pypdf
